+++ packages/backend/src/scheduler/SchedulerTask.ts
@@ -7,11 +7,13 @@ import {
     DownloadCsvPayload,
     EmailNotificationPayload,
     FieldReferenceError,
+    formatRows,
     friendlyName,
     getCustomLabelsFromTableConfig,
     getHiddenTableFields,
     getHumanReadableCronExpression,
     getItemMap,
+    getPivotConfig,
     getRequestMethod,
     getSchedulerUuid,
     GsheetsNotificationPayload,
@@ -31,6 +33,7 @@ import {
     NotificationFrequency,
     NotificationPayloadBase,
     operatorActionValue,
+    pivotResultsAsCsv,
     QueryExecutionContext,
     ScheduledDeliveryPayload,
     SchedulerAndTargets,
@@ -1141,17 +1144,41 @@ export default class SchedulerTask {
                 payload.metricQuery.additionalMetrics,
                 payload.metricQuery.tableCalculations,
             );
-            await this.googleDriveClient.appendToSheet(
-                refreshToken,
-                spreadsheetId,
-                rows,
-                itemMap,
-                payload.showTableNames,
-                undefined, // tabName
-                payload.columnOrder,
-                payload.customLabels,
-                payload.hiddenFields,
-            );
+            if (payload.pivotConfig) {
+                // PivotQueryResults expects a formatted ResultRow[] type, so we need to convert it first
+                // TODO: refactor pivotQueryResults to accept a Record<string, any>[] simple row type for performance
+                const formattedRows = formatRows(rows, itemMap);
+
+                const pivotedResults = pivotResultsAsCsv({
+                    pivotConfig: payload.pivotConfig,
+                    rows: formattedRows,
+                    itemMap,
+                    metricQuery: payload.metricQuery,
+                    customLabels: payload.customLabels,
+                    onlyRaw: true,
+                    maxColumnLimit:
+                        this.lightdashConfig.pivotTable.maxColumnLimit,
+                });
+
+                await this.googleDriveClient.appendCsvToSheet(
+                    refreshToken,
+                    spreadsheetId,
+                    pivotedResults,
+                );
+            } else {
+                await this.googleDriveClient.appendToSheet(
+                    refreshToken,
+                    spreadsheetId,
+                    rows,
+                    itemMap,
+                    payload.showTableNames,
+                    undefined, // tabName
+                    payload.columnOrder,
+                    payload.customLabels,
+                    payload.hiddenFields,
+                );
+            }
+
             const truncated = this.csvService.couldBeTruncated(rows);
 
             await this.schedulerService.logSchedulerJob({
@@ -1613,18 +1640,44 @@ export default class SchedulerTask {
                     undefined,
                     reportUrl,
                 );
+                const pivotConfig = getPivotConfig(chart);
+                if (
+                    pivotConfig &&
+                    isTableChartConfig(chart.chartConfig.config)
+                ) {
+                    // PivotQueryResults expects a formatted ResultRow[] type, so we need to convert it first
+                    // TODO: refactor pivotQueryResults to accept a Record<string, any>[] simple row type for performance
+                    const formattedRows = formatRows(rows, itemMap);
 
-                await this.googleDriveClient.appendToSheet(
-                    refreshToken,
-                    gdriveId,
-                    rows,
-                    itemMap,
-                    showTableNames,
-                    undefined,
-                    chart.tableConfig.columnOrder,
-                    customLabels,
-                    getHiddenTableFields(chart.chartConfig),
-                );
+                    const pivotedResults = pivotResultsAsCsv({
+                        pivotConfig,
+                        rows: formattedRows,
+                        itemMap,
+                        metricQuery: chart.metricQuery,
+                        customLabels,
+                        onlyRaw: true,
+                        maxColumnLimit:
+                            this.lightdashConfig.pivotTable.maxColumnLimit,
+                    });
+
+                    await this.googleDriveClient.appendCsvToSheet(
+                        refreshToken,
+                        gdriveId,
+                        pivotedResults,
+                    );
+                } else {
+                    await this.googleDriveClient.appendToSheet(
+                        refreshToken,
+                        gdriveId,
+                        rows,
+                        itemMap,
+                        showTableNames,
+                        undefined,
+                        chart.tableConfig.columnOrder,
+                        customLabels,
+                        getHiddenTableFields(chart.chartConfig),
+                    );
+                }
             } else if (dashboardUuid) {
                 const dashboard = await this.dashboardService.getById(
                     user,
@@ -1727,18 +1780,46 @@ export default class SchedulerTask {
                                 gdriveId,
                                 chartNames[chartUuid] || chartUuid,
                             );
+                        const pivotConfig = getPivotConfig(chart);
+                        if (
+                            pivotConfig &&
+                            isTableChartConfig(chart.chartConfig.config)
+                        ) {
+                            // PivotQueryResults expects a formatted ResultRow[] type, so we need to convert it first
+                            // TODO: refactor pivotQueryResults to accept a Record<string, any>[] simple row type for performance
+                            const formattedRows = formatRows(rows, itemMap);
 
-                        await this.googleDriveClient.appendToSheet(
-                            refreshToken,
-                            gdriveId,
-                            rows,
-                            itemMap,
-                            showTableNames,
-                            tabName,
-                            chart.tableConfig.columnOrder,
-                            customLabels,
-                            getHiddenTableFields(chart.chartConfig),
-                        );
+                            const pivotedResults = pivotResultsAsCsv({
+                                pivotConfig,
+                                rows: formattedRows,
+                                itemMap,
+                                metricQuery: chart.metricQuery,
+                                customLabels,
+                                onlyRaw: true,
+                                maxColumnLimit:
+                                    this.lightdashConfig.pivotTable
+                                        .maxColumnLimit,
+                            });
+
+                            await this.googleDriveClient.appendCsvToSheet(
+                                refreshToken,
+                                gdriveId,
+                                pivotedResults,
+                                tabName,
+                            );
+                        } else {
+                            await this.googleDriveClient.appendToSheet(
+                                refreshToken,
+                                gdriveId,
+                                rows,
+                                itemMap,
+                                showTableNames,
+                                tabName,
+                                chart.tableConfig.columnOrder,
+                                customLabels,
+                                getHiddenTableFields(chart.chartConfig),
+                            );
+                        }
                     }, Promise.resolve())
                     .catch((error) => {
                         Logger.debug('Error processing charts:', error);